content_engine/requirements.txt
streamlit
PyPDF2
sentence-transformers
faiss-cpu
pygpt4all
numpy
